Hyper Parameter Tweaks
Changed hyper parameters to those recommended by Professor Taylor.
Gamma: 1.0
Learning Rate: 0.1

Agent still does worse than random. I think it is because of how the
agent will continually try to walk into walls.
Going to change GetQValue() (or whatever it is called) in QHat so that
when the agent takes an action that results in no change in state then
it will return a large negative value.